instagram/requirements.txt
asgiref==3.11.1
dj-database-url==3.0.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.0
gunicorn==23.0.0
pillow==12.1.1
psycopg[binary]==3.2.10
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.11.0
